change gateway models
