+++ gateway/internal/app/models/users.go
@@ -1,0 +1,16 @@
+package models
+
+import "time"
+
+type UserProfile struct {
+	UserID    string
+	Nickname  string
+	Bio       string
+	AvatarURL string
+	CreatedAt time.Time
+	UpdatedAt time.Time
+}
+
+type UserSearchResult struct {
+	Profiles []*UserProfile
+}
